Reorganized of standard types
