+++ empyric/__init__.py
@@ -1,7 +1,4 @@
 import argparse
-from abc import ABC
-from numpy import ndarray
-from pandas import Series, DataFrame
 
 try:
     import pytest
@@ -88,33 +85,3 @@ def execute():
     else:
         manager = Manager(runcard=args.runcard)
         manager.run(directory=args.directory)
-
-class Toggle(ABC):
-
-    def __init__(self, on_or_off: str):
-        self.on = True if on_or_off.lower() == 'on' else False
-
-    def __bool__(self):
-        return True if self.on else False
-
-    def __eq__(self, other):
-        return True if self.on == other.on else False
-
-
-ON = Toggle('ON')
-OFF = Toggle('OFF')
-
-
-class Array(ABC):
-    """
-    Abstract base class for all array-like types, essentially any commonly
-    used type that can be indexed
-    """
-    pass
-
-
-Array.register(list)
-Array.register(tuple)
-Array.register(ndarray)
-Array.register(Series)
-Array.register(DataFrame)
