+++ empyric/collection/thermometers.py
@@ -1,15 +1,13 @@
 import importlib
-import time
-import numpy as np
+from empyric.adapters import Phidget, Serial
+from empyric.instruments import Instrument, setter, measurer
+from empyric.types import recast
 
-from empyric.adapters import *
-from empyric.collection.instrument import *
-from empyric.tools import to_number
 
 class Phidget1101(Instrument):
     """
     Phidgets 4x TC reader
-    Many instrumet methods (setX, getY, etc.) are mapped by the adapter from
+    Many instrument methods (setX, getY, etc.) are mapped by the adapter from
     the Phidgets device class
     """
 
@@ -73,14 +71,14 @@ class WilliamsonPyrometer(Instrument):
 
     @measurer
     def measure_temperature(self):
-        # temp returned in in F, convert to C
-        return (to_number(self.query('FT')) - 32) / 1.8
+        # temp returned in F, convert to C
+        return (recast(self.query('FT')) - 32) / 1.8
 
     @measurer
     def measure_unfiltered_temperature(self):
-        # temp returned in in F, convert to C
-        return (to_number(self.query('UT')) - 32) / 1.8
+        # temp returned in F, convert to C
+        return (recast(self.query('UT')) - 32) / 1.8
 
     @measurer
     def measure_signal_strength(self):
-        return to_number(self.query('SS'))
+        return recast(self.query('SS'))
